api/requirements.txt
azure-functions
pandas
azure-storage-blob
azure-identity
pyodbc